removed camel casing from namespaces
