+++ artemis/src/main/java/tech/pegasys/artemis/datastructures/beaconchainblocks/BeaconBlock.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.artemis.datastructures.BeaconChainBlocks;
+package tech.pegasys.artemis.datastructures.beaconchainblocks;
 
 import tech.pegasys.artemis.ethereum.core.Hash;
 import tech.pegasys.artemis.util.uint.UInt384;
