+++ artemis/src/main/java/tech/pegasys/artemis/state/BeaconState.java
@@ -29,14 +29,14 @@ import static tech.pegasys.artemis.Constants.PENDING_ACTIVATION;
 import static tech.pegasys.artemis.Constants.WHISTLEBLOWER_REWARD_QUOTIENT;
 import static tech.pegasys.artemis.ethereum.core.TreeHash.hash_tree_root;
 
-import tech.pegasys.artemis.datastructures.BeaconChainState.CandidatePoWReceiptRootRecord;
-import tech.pegasys.artemis.datastructures.BeaconChainState.CrosslinkRecord;
-import tech.pegasys.artemis.datastructures.BeaconChainState.ForkData;
-import tech.pegasys.artemis.datastructures.BeaconChainState.PendingAttestationRecord;
-import tech.pegasys.artemis.datastructures.BeaconChainState.ShardCommittee;
-import tech.pegasys.artemis.datastructures.BeaconChainState.ShardReassignmentRecord;
-import tech.pegasys.artemis.datastructures.BeaconChainState.ValidatorRecord;
-import tech.pegasys.artemis.datastructures.BeaconChainState.ValidatorRegistryDeltaBlock;
+import tech.pegasys.artemis.datastructures.beaconchainstate.CandidatePoWReceiptRootRecord;
+import tech.pegasys.artemis.datastructures.beaconchainstate.CrosslinkRecord;
+import tech.pegasys.artemis.datastructures.beaconchainstate.ForkData;
+import tech.pegasys.artemis.datastructures.beaconchainstate.PendingAttestationRecord;
+import tech.pegasys.artemis.datastructures.beaconchainstate.ShardCommittee;
+import tech.pegasys.artemis.datastructures.beaconchainstate.ShardReassignmentRecord;
+import tech.pegasys.artemis.datastructures.beaconchainstate.ValidatorRecord;
+import tech.pegasys.artemis.datastructures.beaconchainstate.ValidatorRegistryDeltaBlock;
 import tech.pegasys.artemis.ethereum.core.Hash;
 import tech.pegasys.artemis.util.uint.UInt384;
 import tech.pegasys.artemis.util.uint.UInt64;
